+++ pages/formula.php
@@ -7,16 +7,10 @@ if(mysqli_num_rows(mysqli_query($conn, "SELECT id FROM formulasMetaData WHERE fi
 	echo 'Formula doesn\'t exist';
 	exit;
 }
-$formula_q = mysqli_query($conn, "SELECT * FROM formulas WHERE name = '$f_name' ORDER BY ingredient ASC");
-                    
-
-$mg = mysqli_fetch_assoc(mysqli_query($conn, "SELECT SUM(quantity) AS total_mg FROM formulas WHERE name = '$f_name'"));
-$meta = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM formulasMetaData WHERE name = '$f_name'"));
-
-$top_calc = calcPerc($f_name, 'Top', $settings['top_n'], $conn);
-$heart_calc = calcPerc($f_name, 'Heart', $settings['heart_n'], $conn);
-$base_calc = calcPerc($f_name, 'Base', $settings['base_n'], $conn);
-
+if(mysqli_num_rows(mysqli_query($conn, "SELECT fid FROM formulas WHERE fid = '$fid'"))){
+	$legend = 1;
+}
+$meta = mysqli_fetch_array(mysqli_query($conn, "SELECT id,image FROM formulasMetaData WHERE fid = '$fid'"));
 ?>
 <style>
 .mfp-iframe-holder .mfp-content {
@@ -33,7 +27,7 @@ $base_calc = calcPerc($f_name, 'Base', $settings['base_n'], $conn);
           <div class="card shadow mb-4">
             <div class="card-header py-3"> 
 			  <?php if($meta['image']){?><div class="img-formula"><img class="img-perfume" src="<?php echo $meta['image']; ?>"/></div><?php } ?>
-              <h2 class="m-0 font-weight-bold text-primary"><a href="?do=Formula&name=<?php echo $f_name; ?>"><?php echo $f_name; ?></a></h2>
+              <h2 class="m-0 font-weight-bold text-primary"><a href="?do=Formula&name=<?=base64_decode($fid)?>"><?=base64_decode($fid)?></a></h2>
               <h5 class="m-1 text-primary"><a href="pages/getFormMeta.php?id=<?php echo $meta['id'];?>" class="popup-link">Details</a></h5>
             </div>
         <!-- Nav tabs -->
@@ -41,6 +35,7 @@ $base_calc = calcPerc($f_name, 'Base', $settings['base_n'], $conn);
           <li class="active"><a href="#main_formula" role="tab" data-toggle="tab"><icon class="fa fa-bong"></icon> Formula</a></li>
     	  <li><a href="#impact" role="tab" data-toggle="tab"><i class="fa fa-magic"></i> Notes Impact</a></li>
           <li><a href="#pyramid" role="tab" data-toggle="tab"><i class="fa fa-table"></i> Olfactory Pyramid</a></li>
+          <li><a href="#summary" role="tab" data-toggle="tab"><i class="fa fa-cubes"></i> Notes Summary</a></li>
         </ul>
                      
         <div class="tab-content">
@@ -91,10 +86,12 @@ $base_calc = calcPerc($f_name, 'Base', $settings['base_n'], $conn);
                     	<div class="loader-text"></div>
                 	</div>
                 </div>
-                <div>
+                <?php if($legend){ ?>
+                <div id="legend">
                 <p></p>
                 <p>*Values in: <strong class="alert alert-danger">red</strong> exceeds usage level,   <strong class="alert alert-warning">yellow</strong> have no usage level set,   <strong class="alert alert-success">green</strong> are within usage level</p>
                 </div>
+                <?php } ?>
             </div>
           </div>
         </div>
@@ -111,7 +108,20 @@ $base_calc = calcPerc($f_name, 'Base', $settings['base_n'], $conn);
 		        <div id="fetch_pyramid"><div class="loader"></div></div>
 			</div>            
           </div>
-          
+      
+          <div class="tab-pane fade" id="summary">
+            <div class="card-body">
+		        <div id="fetch_summary"><div class="loader"></div></div>
+                <?php if($legend){ ?>
+                <div id="share">
+               	  <p>To include this page in your web site, copy this line and paste it into your html code</p>
+               	  <p><pre>&lt;iframe src=&quot;<?=$_SERVER['REQUEST_SCHEME']?>://<?=$_SERVER['SERVER_NAME']?>/pages/viewSummary.php?id=<?=$fid?>&quot; title=&quot;<?=base64_decode($fid)?>&quot;&gt;&lt;/iframe&gt;</pre></p>
+                	<p>For documentation and parameterisation please refer to: <a href="https://www.jbparfum.com/knowledge-base/share-formula-notes/" target="_blank">https://www.jbparfum.com/knowledge-base/share-formula-notes/</a></p>
+                </div>
+                <?php } ?>
+			</div>            
+          </div>
+                    
         </div>
        </div>         
      </div><!--tabs-->
@@ -259,5 +269,23 @@ function fetch_impact(){
 		}
 	});
 }
+
 fetch_impact();
+
+function fetch_summary(){
+$.ajax({ 
+    url: 'pages/viewSummary.php', 
+	type: 'get',
+    data: {
+		id: "<?=$fid?>"
+		},
+	dataType: 'html',
+		success: function (data) {
+			$('#fetch_summary').html(data);
+		}
+	});
+}
+
+fetch_summary();
+
 </script>